Refactor: Move UserEntity and update RegisterViewModel to save user
- Moved UserEntity to `me.androidbox.authentication.register.data` package.
- Updated RegisterViewModel to use UserRepository to save the user upon successful registration.
- Adjusted UserEntity in `me.androidbox.user.domain` to have a default id.
- Updated import paths for UserEntity in various files.
