+++ composeApp/src/commonMain/kotlin/me/androidbox/authentication/register/data/UserEntity.kt
@@ -1,4 +1,4 @@
-package me.androidbox.authentication.register
+package me.androidbox.authentication.register.data
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
