+++ composeApp/src/commonMain/kotlin/me/androidbox/core/data/NoteMarkDatabase.kt
@@ -2,7 +2,7 @@ package me.androidbox.core.data
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import me.androidbox.authentication.register.UserEntity
+import me.androidbox.authentication.register.data.UserEntity
 import me.androidbox.notes.data.NoteMarkDao
 import me.androidbox.notes.data.models.NoteItemEntity
