+++ composeApp/src/commonMain/kotlin/me/androidbox/notes/data/NoteMarkDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Delete
 import androidx.room.Query
 import androidx.room.Upsert
 import kotlinx.coroutines.flow.Flow
-import me.androidbox.authentication.register.UserEntity
+import me.androidbox.authentication.register.data.UserEntity
 import me.androidbox.notes.data.models.NoteItemEntity
 
 @Dao
