+++ composeApp/src/commonMain/kotlin/me/androidbox/user/data/UserLocalDataSource.kt
@@ -1,6 +1,6 @@
 package me.androidbox.user.data
 
-import me.androidbox.authentication.register.UserEntity
+import me.androidbox.authentication.register.data.UserEntity
 import me.androidbox.core.models.DataError
 import me.androidbox.user.domain.User
 import net.orandja.either.Either
