+++ composeApp/src/commonMain/kotlin/me/androidbox/user/data/imp/UserLocalDataSourceImp.kt
@@ -1,6 +1,6 @@
 package me.androidbox.user.data.imp
 
-import me.androidbox.authentication.register.UserEntity
+import me.androidbox.authentication.register.data.UserEntity
 import me.androidbox.core.data.NoteMarkDatabase
 import me.androidbox.core.models.DataError
 import me.androidbox.user.data.UserLocalDataSource
