+++ composeApp/src/commonMain/kotlin/me/androidbox/user/domain/UserEntity.kt
@@ -1,7 +1,7 @@
 package me.androidbox.user.domain
 
 data class User(
-    val id: Int,
+    val id: Int = 0,
     val userName: String,
     val email: String
 )
